Extract story parameters into a shared helper and refactor Heading stories
