--- packages/Heading/stories/storyHelpers.js
@@ -1,7 +1,0 @@
-export const exampleStoryParameters = {
-  options: {
-    isToolshown: true,
-    showPanel: false,
-  },
-  viewMode: "story",
-};
